feat(python): break check_file_size.py into a package and update command registration (issue #10)
